streamlit_scripts/requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
altair==4.2.0
async-timeout==4.0.2
asynctest==0.13.0
attrs==22.1.0
backports.zoneinfo==0.2.1
blinker==1.5
cachetools==5.2.0
charset-normalizer==2.1.1
click==8.1.3
commonmark==0.9.1
decorator==5.1.1
entrypoints==0.4
filelock==3.8.0
fire==0.4.0
frozenlist==1.3.3
fsspec==2022.11.0
gitdb==4.0.10
GitPython==3.1.29
huggingface-hub==0.11.1
idna==3.4
importlib-metadata==5.1.0
importlib-resources==5.10.0
Jinja2==3.1.2
jsonschema==4.17.3
lightning-utilities==0.3.0
MarkupSafe==2.1.1
multidict==6.0.2
numpy==1.21.6
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
packaging==21.3
pandas==1.3.5
Pillow==9.3.0
pkgutil_resolve_name==1.3.10
protobuf==3.20.1
pyarrow==10.0.1
pydeck==0.8.0
Pygments==2.13.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.19.2
python-dateutil==2.8.2
pytorch-lightning==1.8.3.post1
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
regex==2022.10.31
requests==2.28.1
rich==12.6.0
semver==2.13.0
six==1.16.0
smmap==5.0.0
streamlit==1.15.1
tensorboardX==2.5.1
termcolor==2.1.1
tokenizers==0.13.2
toml==0.10.2
toolz==0.12.0
torch==1.13.0
torchmetrics==0.10.3
tornado==6.2
tqdm==4.64.1
transformers==4.24.0
typing_extensions==4.4.0
tzdata==2022.6
tzlocal==4.2
urllib3==1.26.13
validators==0.20.0
watchdog==2.1.9
yarl==1.8.1
zipp==3.11.0
